requirements.txt
pytesseract==0.3.8
Pillow==8.4.0
pywhatkit==5.1
tk==0.1.0
